refactor: rename Navigation with Header and make it as a widget
